move iso8601 formatting to utils
