+++ nominal/_timeutils.py
@@ -3,6 +3,7 @@ from __future__ import annotations
 from datetime import datetime, timezone
 
 import dateutil.parser
+import numpy as np
 
 from ._api.combined import ingest_api, scout_run_api
 
@@ -53,3 +54,13 @@ def _parse_timestamp(ts: str | datetime | IntegralNanosecondsUTC) -> IntegralNan
     if isinstance(ts, str):
         ts = dateutil.parser.parse(ts)
     return _datetime_to_integral_nanoseconds(ts)
+
+
+def _flexible_to_iso8601(ts: datetime | IntegralNanosecondsUTC) -> str:
+    """datetime.datetime objects are only microsecond-precise, so we use numpy's datetime64[ns] for nanosecond precision."""
+    if isinstance(ts, datetime):
+        return ts.astimezone(tz=timezone.utc).isoformat()
+    if isinstance(ts, int):
+        # np.datetime64[ns] assumes UTC
+        return str(np.datetime64(ts, "ns")) + "Z"
+    raise TypeError(f"timestamp {ts} must be a datetime or an integer")
